% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateReliability.R
\name{calculateReliability}
\alias{calculateReliability}
\title{calculateReliability: determine the reliability and SEM per Type}
\usage{
calculateReliability(mydata, n)
}
\arguments{
\item{mydata}{A dataframe containing columns ID, Type, Score (numeric)}

\item{n}{A vector containing for each Type the number of score or assessments assessments, e.g. averages, requirements.}
}
\value{
A list containing 2 vectors; one vector with the reliability coefficient of each Type, the other vector with the SEM values for each Type
}
\description{
calculateReliability: determine the reliability and SEM per Type
}
\examples{
rel <- calculateReliability(mydata, n=c("A"=3,"B"=3,C="2"))
}
