% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GStudy.R
\name{GStudy}
\alias{GStudy}
\title{GStudy}
\usage{
GStudy(mydata, nrDigitsOutput = 4, optimizeSEM = TRUE)
}
\arguments{
\item{mydata}{A dataset containing columns ID (int), Type (enum), Score (float)}

\item{nrDigitsOutput}{Integer, number of digits in the output}

\item{optimizeSEM}{Boolean, if TRUE, the weights are adjusted in order to minimize the Standard Error of Measurement (SEM)}
}
\value{
List containing a matrix with descriptive statistics (statisticMatrix) and textual output (outputGstudy) of the composite reliabitiy
#' @export
}
\description{
GStudy
}
\examples{
outputG <- GStudy(mydata,4,TRUE)
outputG$statisticMatrix
outputG$outputGstudy
}
