% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GStudyPerType.R
\name{GStudyPerType}
\alias{GStudyPerType}
\title{GStudyPerType: This function is mainly used within calculateVarCov.R, but can be executed on its own to determine the reliability coefficient and SEM for a dataset with a single type of assessment.}
\usage{
GStudyPerType(dataPerAssessmentType)
}
\arguments{
\item{dataPerAssessmentType}{A dataframe containing columns ID, Type, Score (numeric), with only one value in column Type}
}
\value{
A matrix presenting the observerd varianced and residual, number of ID's and the percentage of the total variance for each group
}
\description{
GStudyPerType: This function is mainly used within calculateVarCov.R, but can be executed on its own to determine the reliability coefficient and SEM for a dataset with a single type of assessment.
}
