% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{mydata}
\format{
\subsection{\code{mydata}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{ID}{ID of the student}
\item{Type}{The type of assessment}
\item{Score}{The obtained score by this student on this occasion, using the type of assessment}
...
}
}
}
\usage{
mydata
}
\description{
A dataset that can be used as example in package CompositeReliability.
}
\keyword{datasets}
