% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DStudy.R
\name{DStudy}
\alias{DStudy}
\title{DStudy: the program presents the reliability coefficient and the SEM for different numbers of assessments per type.
Both the reliability coefficient and the SEM are presented in graphs for differing numbers of assessments,
given insight in the impact on the reliability if more or less assessments per type were required or advised.}
\usage{
DStudy(mydata, maxNrAssessments = 60)
}
\arguments{
\item{mydata}{A dataframe containing columns ID, Type, Score (numeric)}

\item{maxNrAssessments}{The maximum (Int) number of assessments per type on with the D study is executed}
}
\value{
A list containing 2 plots: reliability (plotRel) and Standard Error of Measurement SEM (plotSEM)
}
\description{
DStudy: the program presents the reliability coefficient and the SEM for different numbers of assessments per type.
Both the reliability coefficient and the SEM are presented in graphs for differing numbers of assessments,
given insight in the impact on the reliability if more or less assessments per type were required or advised.
}
\examples{
plots <- DStudy(mydata, maxNrAssessments = 10)
}
