% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GStudy.R
\name{GStudy}
\alias{GStudy}
\title{GStudy for a dataset in which every student p has a potentially differing number of scores i on each assessment type m.
i.e. model i: (p x m)}
\usage{
GStudy(mydata, nrDigitsOutput = 4, optimizeSEM = TRUE)
}
\arguments{
\item{mydata}{A dataset containing columns ID (int), Type (enum), Score (float)}

\item{nrDigitsOutput}{Integer, number of digits in the output}

\item{optimizeSEM}{Boolean, if TRUE, the weights are adjusted in order to minimize the Standard Error of Measurement (SEM)}
}
\value{
List containing a matrix with descriptive statistics (statisticMatrix) for each Type, and textual output (outputGstudy) of the composite reliabitiy
}
\description{
GStudy for a dataset in which every student p has a potentially differing number of scores i on each assessment type m.
i.e. model i: (p x m)
}
\examples{
outputG <- GStudy(mydata,nrDigitsOutput=4,optimizeSEM=TRUE)
outputG$statisticMatrix
outputG$outputGstudy
}
