% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GStudy.R
\name{GStudy}
\alias{GStudy}
\title{GStudy for a dataset in which every student p has a potentially differing number of scores i on each assessment type m.
i.e. model i: (p x m). The output gives descriptive statistics, reliability coefficient and SEM for each assessment type.}
\usage{
GStudy(mydata, nrDigitsOutput = 4)
}
\arguments{
\item{mydata}{A dataframe containing columns ID, Type, Score (numeric)}

\item{nrDigitsOutput}{Integer, number of digits in the output}
}
\value{
Matrix with descriptive statistics for each Type of assessment
}
\description{
GStudy for a dataset in which every student p has a potentially differing number of scores i on each assessment type m.
i.e. model i: (p x m). The output gives descriptive statistics, reliability coefficient and SEM for each assessment type.
}
\examples{
GStudy(mydata,nrDigitsOutput=4)
}
